Kick off the initial timeline doc 🕰️
[SIG interoperability](https://github.com/cdfoundation/sig-interoperability)
started working on a white paper, and found themselves wanting to define
terms they used such as "continuous delivery" and "CI/CD". The call to
define these terms brought so much feedback that it has been split into
[a separate document](https://docs.google.com/document/d/1IUSmgtw5eC2JwyfiX7IPK3twl3MSjWAsmCKmJ6Y5z-w/edit#)
in which we are continuing to iterate on these definitions.

In order to help move this discussion forward, I think it's useful to be
able to look back at the history of these terms and how they came into
being.

The document added in this commit lists some of the impactful sources of
these terms.

I've gone back and forth on what to include, and whether or not it makes
sense to identify some of these sources as more impactful than others
(i.e. "canonical"). I also initially included every blog post and book I
encountered on this topic and am still not sure if it makes sense to
track that here. I'm leaning towards including it, but maybe in a
separate table or doc?

Anyway I want to keep adding to this! Via
[this blog post by @jezhumble from 2010](https://continuousdelivery.com/2010/02/continuous-delivery/)
I've discovered that the phrase "continuous delivery" was inspired by
the principles behind the Agile Manifesto, so I want to include the
Agile Manifesto in this list as well, and I want to include more about
Continuous integration (discovered via
[a series of blog posts by circleci](https://circleci.com/blog/a-brief-history-of-devops-part-iii-automated-testing-and-continuous-integration/)).

And I also discovered via
[a @jezhumble blog post](https://continuousdelivery.com/2010/08/continuous-delivery-vs-continuous-deployment/)
that "Continuous Deployment" actually came first which was fascinating!!

Anyway I hope this is a good starting point!!