Add initial definitions for CD terms 📙
The discussion about these definitions was kicked off in the
interoperability working group, then moved to
[this google doc](https://docs.google.com/document/d/1IUSmgtw5eC2JwyfiX7IPK3twl3MSjWAsmCKmJ6Y5z-w/edit#)
and now is being moved to this repo, after having established (at least
an initial) timeline in timeline.md of when these terms came into being.

This is intended to just be the first iteration of these defintions and
to be refined over time.